removing comments, reorganizing file structure and changed folder names
